Only use ES5 in distributed bundle (#6399)
* IE11: Don't use let in final bundle

* Make benchmark bundle wrapper match main bundle

* Extract bundle prelude into separate file

* Test that distributed bundle contains only ES5
