+++ bench/benchmarks.js
@@ -42,9 +42,14 @@ register(FilterCreate);
 register(FilterEvaluate);
 
 import getWorkerPool from '../src/util/global_worker_pool';
-// Set up the worker blob URL--written to window.mapboxGlWorkerUrl before this
-// module is executed--before acquiring workers.
-mapboxgl.workerUrl = window.mapboxGlWorkerUrl;
-// Ensure the global worker pool is never drained. Browsers have resource limits
-// on the max number of workers that can be created per page.
-getWorkerPool().acquire(-1);
+
+setTimeout(() => {
+    // Ensure the global worker pool is never drained. Browsers have resource limits
+    // on the max number of workers that can be created per page.
+    // We do this async to avoid creating workers before the worker bundle blob
+    // URL has been set up, which happens after this module is executed.
+    getWorkerPool().acquire(-1);
+}, 0);
+
+export default mapboxgl;
+
